+++ tests/testthat/test_fmt_luminex_results.R
@@ -1,0 +1,4 @@
+library(snprcspf)
+context("fmt_luminex_results")
+test_that("fmt_luminex_results saves appropriate file contents", {
+})